organized simulation folder
