--- simulations/cyl_test/problem_aneu_resistance.py
@@ -1,285 +1,0 @@
-from dolfin import *
-import os
-from turtleFSI.problems import *
-import numpy as np
-from os import path, makedirs, getcwd
-
-# set compiler arguments
-parameters["form_compiler"]["quadrature_degree"] = 6 # Not investigated thorougly. See MSc theses of Gjertsen. 
-parameters["form_compiler"]["optimize"] = True
-# The "ghost_mode" has to do with the assembly of form containing the facet
-# normals n('+') within interior boundaries (dS). for 3D mesh the value should
-# be "shared_vertex", for 2D mesh "shared_facet", the default value is "none"
-parameters["ghost_mode"] = "shared_vertex" 
-_compiler_parameters = dict(parameters["form_compiler"])
-
-
-def set_problem_parameters(default_variables, **namespace):
-    # Overwrite default values
-    E_s_val = 1E6  # Young modulus (Pa)
-    nu_s_val = 0.45
-    mu_s_val = E_s_val / (2 * (1 + nu_s_val))  # 0.345E6
-    lambda_s_val = nu_s_val * 2. * mu_s_val / (1. - 2. * nu_s_val)
-
-    default_variables.update(dict(
-        T=0.2, # Simulation end time
-        dt=0.001, # Timne step size
-        theta=0.501, # Theta scheme (implicit/explicit time stepping)
-        atol=1e-5, # Absolute tolerance in the Newton solver
-        rtol=1e-5,# Relative tolerance in the Newton solver
-        mesh_file="artery_coarse_rescaled", # duh
-        inlet_id=2,  # inlet id 
-        outlet_id1=3,  # outlet nb
-        inlet_outlet_s_id=11,  # also the "rigid wall" id for the stucture problem. MB: Clarify :)
-        fsi_id=22,  # fsi Interface 
-        rigid_id=11,  # "rigid wall" id for the fluid and mesh problem
-        outer_wall_id=33,  # outer surface / external id 
-        rho_f=1.025E3,    # Fluid density [kg/m3]
-        mu_f=3.5E-3,       # Fluid dynamic viscosity [Pa.s]
-        rho_s=1.0E3,    # Solid density [kg/m3]
-        mu_s=mu_s_val,     # Solid shear modulus or 2nd Lame Coef. [Pa]
-        nu_s=nu_s_val,      # Solid Poisson ratio [-]
-        lambda_s=lambda_s_val,  # Solid 1rst Lamé coef. [Pa]
-#        dx_f_id=1,      # ID of marker in the fluid domain. When reading the mesh, the fuid domain is assigned with a 1. Old crap :)
-#        dx_s_id=2,      # ID of marker in the solid domain
-        extrapolation="laplace",  # laplace, elastic, biharmonic, no-extrapolation
-        # ["constant", "small_constant", "volume", "volume_change", "bc1", "bc2"]
-        extrapolation_sub_type="constant",
-        recompute=3000, # Number of iterations before recompute Jacobian. 
-        recompute_tstep=1, # Number of time steps before recompute Jacobian. 
-        save_step=1, # Save frequency of files for visualisation
-        folder="Resistance_BC",# 
-        checkpoint_step=5, # 
-        save_deg=1 # 1 saves the nodal values only while 2 takes full advantage of the mide side nodes available in the P2 solution. P2 f0r nice visualisations :)
-        # probe point(s)
-    ))
-
-    return default_variables
-
-
-def get_mesh_domain_and_boundaries(mesh_file, fsi_id, rigid_id, outer_wall_id, folder, **namespace):
-    print("Obtaining mesh, domains and boundaries...")
-    mesh = Mesh()
-    hdf = HDF5File(mesh.mpi_comm(), "mesh/" + mesh_file + ".h5", "r")
-    hdf.read(mesh, "/mesh", False)
-    boundaries = MeshFunction("size_t", mesh, 2)
-    hdf.read(boundaries, "/boundaries")
-    domains = MeshFunction("size_t", mesh, 3)
-    hdf.read(domains, "/domains")
-
-    # Only considere FSI in domain within this sphere # Corrected sphere for the rescaled artery!
-    sph_x = 1.8124701455235481e-06
-    sph_y = -6.299931555986404e-06
-    sph_z = -7.300404831767082e-07
-    sph_rad = 0.0034
-    print("The coordinates of the FSI sphere are [{}, {}, {}] of radius:{}".format(sph_x, sph_y, sph_z, sph_rad))
-
-    i = 0
-    for submesh_facet in facets(mesh):
-        idx_facet = boundaries.array()[i]
-        if idx_facet == fsi_id:
-            vert = submesh_facet.entities(0)
-            mid = submesh_facet.midpoint()
-            dist_sph_center = sqrt((mid.x() - sph_x)**2 + (mid.y() - sph_y)**2 + (mid.z() - sph_z)**2)
-            if dist_sph_center > sph_rad:
-                boundaries.array()[i] = rigid_id
-        if idx_facet == outer_wall_id:
-            vert = submesh_facet.entities(0)
-            mid = submesh_facet.midpoint()
-            dist_sph_center = sqrt((mid.x() - sph_x)**2 + (mid.y() - sph_y)**2 + (mid.z() - sph_z)**2)
-            if dist_sph_center > sph_rad:
-                boundaries.array()[i] = rigid_id
-        i += 1
-    print("Obtained mesh, domains and boundaries.")
-
-    return mesh, domains, boundaries
-
-
-class StokesPrb():
-
-    def __init__(self, mesh, domains, boundaries, dsi, inlet_id, outlet_id1, normal, mu_f):
-
-        self.mesh = mesh
-        self.ve = VectorElement('CG', self.mesh.ufl_cell(), 2)
-        self.pe = FiniteElement('CG', self.mesh.ufl_cell(), 1)
-        self.Elem = MixedElement([self.ve, self.pe])
-        self.VP = FunctionSpace(self.mesh, self.Elem)
-
-        self.u_inflow_exp = VelInPara(t=0.0, t_ramp=0.0, n=normal, dsi=dsi, mesh=mesh, degree=3)
-        self.u_inlet = DirichletBC(self.VP.sub(0), self.u_inflow_exp, boundaries, inlet_id)
-        self.u_fsi1 = DirichletBC(self.VP.sub(0), ((0.0, 0.0, 0.0)), boundaries, 11)
-        self.u_fsi2 = DirichletBC(self.VP.sub(0), ((0.0, 0.0, 0.0)), boundaries, 22)
-        self.u_fsi3 = DirichletBC(self.VP.sub(0), ((0.0, 0.0, 0.0)), boundaries, 33)
-
-        self.Pout_val = Constant(0.0)
-
-        self.bcs = [self.u_inlet, self.u_fsi1, self.u_fsi2, self.u_fsi3]
-
-        self.ds = Measure("ds", domain=self.mesh, subdomain_data=boundaries)
-        self.n = FacetNormal(self.mesh)
-        # Define variational problem
-        (self.u, self.p) = TrialFunctions(self.VP)
-        (self.v, self.q) = TestFunctions(self.VP)
-        f = Constant((0.0, 0.0, 0.0))
-        self.a = (mu_f*inner(grad(self.u), grad(self.v))*dx
-                  - self.p*div(self.v)*dx
-                  + self.q*div(self.u)*dx)
-        self.L = (inner(f, self.v)*dx
-                  - self.Pout_val * inner(self.n, self.v)*self.ds(outlet_id1))
-
-    def solve(self):
-
-        # Solution vector
-        self.w = Function(self.VP)
-        # Assemble system and Solve
-        A, B = assemble_system(self.a, self.L, self.bcs)
-        #solve(A, self.w.vector(), B, "minres", "amg")
-        loc_sol = LUSolver(Matrix(), "mumps")
-        loc_sol.solve(A, self.w.vector(), B)
-        # Get sub-functions
-        self.u, self.p = self.w.split(deepcopy=True)
-
-        return self.u, self.p
-
-
-class VelInPara(UserExpression):
-    def __init__(self, t, t_ramp, n, dsi, mesh, **kwargs):
-        self.t = t
-        self.t1 = t_ramp
-        self.n = n
-        self.dsi = dsi
-        self.d = mesh.geometry().dim()
-        self.x = SpatialCoordinate(mesh)
-        # Compute area of boundary tesselation by integrating 1.0 over all facets
-        self.A = assemble(Constant(1.0, name="one")*self.dsi)
-        # Compute barycenter by integrating x components over all facets
-        self.c = [assemble(self.x[i]*self.dsi) / self.A for i in range(self.d)]
-        # Compute radius by taking max radius of boundary points
-        self.r = np.sqrt(self.A / np.pi)
-        super().__init__(**kwargs)
-
-    def update(self, t):
-        self.t = t
-        
-    def eval(self, value, x):
-        if self.t < 0.1:
-            v_local_max = self.t*(0.74/0.1) # 0.77 m/s mmHg rise in 0.1s
-        else:
-            v_local_max = self.t*0.9666667 + 0.64333333 # lower slope
-
-        r2 = (x[0]-self.c[0])**2 + (x[1]-self.c[1])**2 + (x[2]-self.c[2])**2  # radius**2
-        fact_r = 1 - (r2/self.r**2)
-
-        value[0] = -self.n[0] * (v_local_max) *fact_r  # *self.t
-        value[1] = -self.n[1] * (v_local_max) *fact_r  # *self.t
-        value[2] = -self.n[2] * (v_local_max) *fact_r  # *self.t
-        #print("X: {}, Y: {}, Z: {}".format(value[0],value[1],value[2]))
-
-    def value_shape(self):
-        return (3,)
-
-
-class InnerP(UserExpression):
-    def __init__(self, t, t_ramp, n, u, dsi, resistance, p_0, **kwargs):
-        self.t = t
-        self.t_ramp = t_ramp
-        self.n = n # Mesh Normal 
-        self.u = u
-        self.p_0 = p_0 # Initial pressure
-        self.P = p_0 # Apply initial pressure to inner pressure variable
-        self.dsi = dsi
-        self.resistance = resistance
-        super().__init__(**kwargs)
-
-    def update(self, t):
-        self.t = t
-
-        # Sanity check numbers
-        diam = 0.005 # metres
-        V_init = 0.77 # m/s
-        Area_Inlet = (3.14159/4)*diam**2
-        p_final_ramp = self.resistance*(V_init/2)*Area_Inlet    
-
-
-        # Caclulate flow rate
-        Q = np.abs(assemble(inner(self.u,self.n) * self.dsi) )
-
-        if MPI.rank(MPI.comm_world) == 0:
-            print("Q = {} m^3/s".format(Q))
-
-        # apply a sigmoid ramp to the pressure 
-        if self.t < self.t_ramp:
-            ramp_factor = (-1/2)*np.cos(3.14159*self.t/self.t_ramp) + 1/2
-        else:
-            ramp_factor = 1.0
-
-        # Caclulate P as resistance boundary condition multiplied by ramp factor
-        self.P = ramp_factor * (self.p_0 + self.resistance * Q)
-        if MPI.rank(MPI.comm_world) == 0:
-            print("P = {} Pa".format(self.P))
-
-    def eval(self, value, x):
-        value[0] = self.P 
-
-    def value_shape(self):
-        return ()
-
-
-def create_bcs(dvp_, DVP, mesh, boundaries, domains, mu_f,
-               fsi_id, outlet_id1, inlet_id, inlet_outlet_s_id,
-               rigid_id, psi, F_solid_linear,p_deg, **namespace):
-
-    if MPI.rank(MPI.comm_world) == 0:
-        print("Create bcs")
-
-
-    dSS = Measure("dS", domain=mesh, subdomain_data=boundaries)
-    n = FacetNormal(mesh)
-
-    # Fluid velocity BCs
-    dsi = ds(inlet_id, domain=mesh, subdomain_data=boundaries)
-    n = FacetNormal(mesh)
-    ndim = mesh.geometry().dim()
-    ni = np.array([assemble(n[i]*dsi) for i in range(ndim)])
-    n_len = np.sqrt(sum([ni[i]**2 for i in range(ndim)]))  # Should always be 1!?
-    normal = ni/n_len
-
-    # new Parabolic profile
-    u_inflow_exp = VelInPara(t=0.0, t_ramp=0.0, n=normal, dsi=dsi, mesh=mesh, degree=3)
-    u_inlet = DirichletBC(DVP.sub(1), u_inflow_exp, boundaries, inlet_id)
-    u_inlet_s = DirichletBC(DVP.sub(1), ((0.0, 0.0, 0.0)), boundaries, inlet_outlet_s_id)
-
-    # Solid Displacement BCs
-    d_inlet = DirichletBC(DVP.sub(0), ((0.0, 0.0, 0.0)), boundaries, inlet_id)
-    d_inlet_s = DirichletBC(DVP.sub(0), ((0.0, 0.0, 0.0)), boundaries, inlet_outlet_s_id)
-    d_rigid = DirichletBC(DVP.sub(0), ((0.0, 0.0, 0.0)), boundaries, rigid_id)
-
-    # Assemble boundary conditions
-    bcs = [u_inlet, d_inlet,  u_inlet_s, d_inlet_s, d_rigid]
-
-    # Solve Stokes problem
-    Stokes = StokesPrb(mesh, domains, boundaries, dsi, inlet_id, outlet_id1, normal, mu_f)
-    sol = Stokes.solve()
-    assign(dvp_["n"].sub(1), sol[0])
-    assign(dvp_["n"].sub(2), sol[1])
-    assign(dvp_["n-1"].sub(1), sol[0])
-    assign(dvp_["n-1"].sub(2), sol[1])
-    assign(dvp_["n-2"].sub(1), sol[0])
-    assign(dvp_["n-2"].sub(2), sol[1])
-
-    # Define the pressure condition (apply to inner surface, numerical instability results from applying to outlet, or using outlet flow rate)
-    p_out_bc_val = InnerP(t=0.0, t_ramp=0.02, n=n, u=sol[0], dsi=dsi, resistance=1e10, p_0=0.0, degree=p_deg)
-    F_solid_linear += p_out_bc_val * inner(n('+'), psi('+'))*dSS(fsi_id)  # defined on the reference domain
-
-
-    return dict(bcs=bcs, u_inflow_exp=u_inflow_exp, p_out_bc_val=p_out_bc_val,
-                F_solid_linear=F_solid_linear)
-
-
-def pre_solve(t, v_, u_inflow_exp, p_out_bc_val, **namespace):
-    # Update the time variable used for the inlet boundary condition
-    u_inflow_exp.update(t)
-    p_out_bc_val.u = v_["n"]
-    p_out_bc_val.update(t)
-
-    return dict(u_inflow_exp=u_inflow_exp, p_out_bc_val=p_out_bc_val)
