--- simulations/stenosis_FC/FC_MCA_10
@@ -1,22 +1,0 @@
-# period_ms = 951
-1	0
--0.03885602	0.389033342
--0.147282143	0.130334152
--0.11754038	0.003512608
--0.020021036	0.009457453
--0.075084472	0.009008134
--0.02801715	-0.043570018
-0.00135982	-0.019541939
-0.005021068	-0.014658254
-0.014010913	-0.000665383
-0.001818315	0.008984162
--0.002389044	0.003027716
--0.002902271	0.002916296
--0.00375792	-0.00072257
--0.000742721	-0.001574683
--0.000450886	-0.000799589
-0.000585817	-0.001015015
-0.000751712	0.000426051
--2.19157E-05	0.000342771
-1.59484E-06	0.000451814
--0.000597902	0.000239054
