--- simulations/stenosis_FC/offset_stenosis.config
@@ -1,20 +1,0 @@
-# user inputs
-case_path="$PROJECT/Aneurysm_Workflow_FSI/simulations/stenosis_FC"
-mesh_path="file_stenosis"
-T=0.951
-dt=0.001
-Q_mean=2.5E-06
-save_deg_sim=1
-save_deg=1 # used in postprocessing, not the simulation
-stride_sd1=1
-stride_sd2=2
-start_t_sd1=0.0
-start_t_sd2=0.0
-end_t_sd1=0.951
-end_t_sd2=0.951
-bands="25,1402"
-x_sphere=0.003
-y_sphere=0.0
-z_sphere=0.0
-r_sphere=0.01
-bands="25,100000"
