--- simulations/stenosis_test/offset_stenosis.config
@@ -1,18 +1,0 @@
-# user inputs
-case_path="stenosis_test"
-mesh_path="file_stenosis"
-refined_mesh_path=$mesh_path"_refined"
-T=0.951
-end_t=$T
-dt=0.001
-Q_mean=6.0E-06
-save_deg_sim=2
-stride_sd1=1
-stride_sd2=2
-start_t_sd1=0.0
-start_t_sd2=0.0
-x_sphere=0.003
-y_sphere=0.0
-z_sphere=0.0
-r_sphere=0.01
-bands="25,100000"
